feat: 검증된 UserDecision 모델 기능 구현
